Killing and Respawning (Part 1 of 2)

+++ CSGO Remake/Assets/Scripts/PlayerWeapons.cs
@@ -5,7 +5,7 @@ public class PlayerWeapons{
 
 
     public string name = "Glock 18";
-    public float damage = 20f;
+    public int damage = 20;
     public int range = 100;
